+++ Orbital/src/loader/gltf/worker.rs
@@ -1,0 +1,11 @@
+use std::thread::JoinHandle;
+
+use crossbeam_channel::Receiver;
+
+use crate::{error::Error, game::WorldChange};
+
+#[derive(Debug)]
+pub struct GLTFWorker {
+    pub receiver: Receiver<Result<Vec<WorldChange>, Error>>,
+    pub worker: JoinHandle<()>,
+}
