+++ Orbital/src/loader/mod.rs
@@ -3,7 +3,7 @@ pub use gltf::*;
 
 use crate::{error::Error, game::WorldChange};
 
-pub trait Loader {
+pub trait Worker {
     fn begin_processing(&mut self);
     fn is_done_processing(&self) -> bool;
     fn finish_processing(&mut self) -> Result<Vec<WorldChange>, Error>;
